do not distinguish between unit and integration tests (GDEV-1389)
The difference between unit and integration tests is very vaquely
defined and only describe two extremes of a spectrum. Thus it is often
not clear where a test belongs. In our current repos we see therefore a
lot of inconsistency in the use of these terms. Thus I would propose to
not differenciate between unit and integration tests for now.

Later we can define a descriptive set of terms to classify our tests
(e.g. "API conformance tests").